Optimize code and improve deployment
Refactor and optimize code for better performance and maintainability.

* **Form Component**: Merge functionality from `FormHandler.vue` into `Form.vue`. Update form submission and error handling methods.
* **StartLoginView.vue**: Replace `FormHandler` with `Form` component. Update form submission and error handling methods.
* **StartSignupView.vue**: Replace `FormHandler` with `Form` component. Update form submission and error handling methods.
* **UserCreateView.vue**: Replace `FormHandler` with `Form` component. Update form submission and error handling methods.
* **UserJoinView.vue**: Replace `FormHandler` with `Form` component. Update form submission and error handling methods.
* **WorkCreateView.vue**: Update form submission to use the merged `Form.vue` component.
* **WorkSettingsView.vue**: Update form submission to use the merged `Form.vue` component. Improve error handling and display methods.
* **vercel.json**: Optimize configuration for better performance and deployment efficiency. Add memory and maxDuration settings for functions.
