+++ client/src/views/start/StartLoginView.vue
@@ -10,7 +10,7 @@
 
   // UI elements
   import Button from '@/components/Button.vue'
-  import FormHandler from '@/components/FormHandler.vue'
+  import Form from '@/components/Form.vue'
 
   onMounted(() => {
     document.title = t('title.login')
@@ -56,7 +56,7 @@
       </p>
     </article>
 
-    <FormHandler
+    <Form
       :requestURL="requestURL"
       :requestData="requestData"
       ref="formComponent"
@@ -103,7 +103,7 @@
           {{ $t('button.forgot-password') }}
         </Button>
       </div>
-    </FormHandler>
+    </Form>
 
     <article>
       <p class="subtext" v-html="$t('text.paragraph.policies')"></p>
