+++ client/src/views/user/UserCreateView.vue
@@ -12,7 +12,7 @@
 
   // UI elements
   import Button from '@/components/Button.vue'
-  import FormHandler from '@/components/FormHandler.vue'
+  import Form from '@/components/Form.vue'
 
   onMounted(() => {
     document.title = t('title.user-create')
@@ -58,7 +58,7 @@
       </p>
     </article>
 
-    <FormHandler
+    <Form
       :requestURL="requestURL"
       :requestData="requestData"
       ref="formComponent"
@@ -127,6 +127,6 @@
           {{ $t('button.submit') }}
         </Button>
       </div>
-    </FormHandler>
+    </Form>
   </main>
 </template>
