+++ vercel.json
@@ -9,5 +9,9 @@
           "destination": "/index.html"
         }
     ],
-    "framework": "vite"
+    "framework": "vite",
+    "functions": {
+        "memory": 1024,
+        "maxDuration": 10
+    }
 }
